requirements.txt
pandas==1.1.0
python-dateutil==2.8.1
requests==2.24.0